+++ Gemfile
@@ -18,10 +18,6 @@ gem 'devise'
 gem 'geocoder'
 gem 'devise_invitable', '~> 2.0.0'
 gem "algoliasearch-rails"
-gem 'google_places'
-
-
-gem 'dotenv-rails', groups: [:development, :test]
 gem 'cloudinary', '~> 1.9.1'
 gem 'carrierwave', '~> 1.2'
